pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "langgraph-agent-deployment"
version = "0.1.0"
authors = [
  { name="Ahmad Arsalan Khateeb", email="ahmadarsalan.khateeb@gmail.com" },
]
description = "A prototyype of deplyment of a langgraph agent."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "fastapi[standard]",
    "uvicorn[standard]",
    "pydantic",
    "langchain",
    "langgraph",
    "langchain_google_genai",
    "python-dotenv",
    "ipykernel",
    "google",
    "ruff",
    "mypy",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
exclude = [".git", "__pycache__", "docs", "venv", "build", "dist"]
select = ["E", "W", "F", "C", "N", "D"]
ignore = ["E203", "E501"]

[tool.mypy]
strict = true
disallow_untyped_defs = true
warn_return_any = true
ignore_missing_imports = false
exclude = ['setup.py', 'docs', 'venv']